Add explicit read status attribute to book
Previously, the read status was defined in the book's tag list. This
meant that users can remove the read status tag, which might break the
(planned) read status filter and read status button. This added
complexity to handle such cases was not worth the effort, compared to
adding a new enum attribute to explicitly define the book's read status.